requirements.txt
certifi==2020.4.5.1
Django==2.1.5
djangorestframework==3.11.0
pytz==2020.1
